Move $FILTERED_IPS before 'allow:any'
Per Anders Savill at Platform.sh:

```All platform http services are backed by nginx and follow the same nginx blocklist rule process which is top to bottom. Once the IP matches a rule it is applied and the rest of the blocklist is not used. As your first rule on the list is 0.0.0.0/0 allow all IPs match this rule and the rest of the rules do not get processed.

If you move the allow rule to the bottom of the list you should see the desired effect.```